add e2e test playwright POC A11y
