+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "alfresco-content-app",
-  "version": "7.3.0",
+  "version": "8.1.0",
   "license": "LGPL-3.0",
   "scripts": {
     "start": "nx serve content-ce",
@@ -28,11 +28,13 @@
     "ci:build": "nx build",
     "ci:build:many": "nx run-many --target=build",
     "ci:playwright:install": "playwright install chromium",
+    "ci:playwright:install:all": "playwright install chromium chrome firefox webkit msedge",
     "ci:e2e": "nx run $E2E_TARGET-e2e:e2e",
     "ci:audit": "adf-cli audit -d docs/audit",
     "ci:licenses": "adf-cli licenses -d docs/licences",
     "ci:changelog": "adf-cli changelog -o docs/changelog --exclude=\"bot\\|Alfresco Build User\\|alfresco-build\"",
-    "ci:check-env": "adf-cli check-cs-env --host $BASE_URL -u $ADMIN_EMAIL -p $ADMIN_PASSWORD"
+    "ci:check-env": "adf-cli check-cs-env --host $BASE_URL -u $ADMIN_EMAIL -p $ADMIN_PASSWORD",
+    "ci:check-search-indexing": "node --env-file-if-exists=.env ./scripts/ci/check-env/check-search-indexing.mjs"
   },
   "repository": {
     "type": "git",
@@ -40,99 +42,97 @@
   },
   "private": true,
   "dependencies": {
-    "@alfresco/adf-content-services": "8.2.1",
-    "@alfresco/adf-core": "8.2.1",
-    "@alfresco/adf-extensions": "8.2.1",
-    "@alfresco/eslint-plugin-eslint-angular": "8.2.1",
-    "@alfresco/js-api": "9.2.1",
-    "@angular/animations": "19.2.6",
-    "@angular/cdk": "19.2.9",
-    "@angular/common": "19.2.6",
-    "@angular/compiler": "19.2.6",
-    "@angular/core": "19.2.6",
-    "@angular/forms": "19.2.6",
-    "@angular/material": "19.2.9",
-    "@angular/material-date-fns-adapter": "19.2.9",
-    "@angular/platform-browser": "19.2.6",
-    "@angular/platform-browser-dynamic": "19.2.6",
-    "@angular/router": "19.2.6",
+    "@alfresco/adf-content-services": "9.2.0-30345919006",
+    "@alfresco/adf-core": "9.2.0-30345919006",
+    "@alfresco/adf-extensions": "9.2.0-30345919006",
+    "@alfresco/eslint-plugin-eslint-angular": "9.2.0-30345919006",
+    "@alfresco/js-api": "10.2.0-30345919006",
+    "@angular/animations": "20.3.25",
+    "@angular/cdk": "20.2.14",
+    "@angular/common": "20.3.25",
+    "@angular/compiler": "20.3.25",
+    "@angular/core": "20.3.25",
+    "@angular/forms": "20.3.25",
+    "@angular/material": "20.2.14",
+    "@angular/material-date-fns-adapter": "20.2.14",
+    "@angular/platform-browser": "20.3.25",
+    "@angular/platform-browser-dynamic": "20.3.25",
+    "@angular/router": "20.3.25",
     "@fontsource/open-sans": "^5.1.0",
-    "@mat-datetimepicker/core": "15.0.2",
-    "@ngrx/effects": "19.2.1",
-    "@ngrx/operators": "19.2.1",
-    "@ngrx/router-store": "19.2.1",
-    "@ngrx/store": "19.2.1",
-    "@ngrx/store-devtools": "19.2.1",
-    "@ngx-translate/core": "^16.0.4",
+    "@mat-datetimepicker/core": "16.0.1",
+    "@ngrx/effects": "~20.1.0",
+    "@ngrx/operators": "~20.1.0",
+    "@ngrx/router-store": "~20.1.0",
+    "@ngrx/store": "~20.1.0",
+    "@ngrx/store-devtools": "~20.1.0",
+    "@ngx-translate/core": "^17.0.0",
     "date-fns": "^4.1.0",
-    "katex": "^0.16.21",
     "material-icons": "^1.13.12",
-    "mermaid": "^11.10.1",
-    "minimatch-browser": "^1.0.0",
-    "ngx-markdown": "19.1.1",
-    "pdfjs-dist": "^5.1.91",
-    "prismjs": "^1.30.0",
+    "pdfjs-dist": "5.1.91",
     "rxjs": "7.8.2",
     "tslib": "2.8.1",
     "zone.js": "0.15.0"
   },
+  "overrides": {
+    "@module-federation/sdk": {
+      "node-fetch": "^2.7.0"
+    }
+  },
   "devDependencies": {
-    "@alfresco/adf-cli": "8.2.1",
-    "@angular-devkit/build-angular": "19.2.15",
-    "@angular-devkit/core": "19.2.7",
-    "@angular-devkit/schematics": "19.2.7",
-    "@angular-eslint/builder": "19.3.0",
-    "@angular-eslint/eslint-plugin": "19.3.0",
-    "@angular-eslint/eslint-plugin-template": "19.3.0",
-    "@angular-eslint/schematics": "19.3.0",
-    "@angular-eslint/template-parser": "19.3.0",
-    "@angular/compiler-cli": "19.2.6",
-    "@angular/language-service": "19.2.6",
-    "@cspell/eslint-plugin": "^9.2.1",
-    "@nx/angular": "22.1.0",
-    "@nx/eslint-plugin": "21.3.1",
-    "@nx/workspace": "21.4.1",
-    "@playwright/test": "^1.56.1",
-    "@reportportal/agent-js-playwright": "^5.1.9",
+    "@alfresco/adf-cli": "9.2.0-30345919006",
+    "@angular-devkit/build-angular": "20.3.30",
+    "@angular-devkit/core": "20.3.16",
+    "@angular-devkit/schematics": "20.3.16",
+    "@angular-eslint/builder": "20.7.0",
+    "@angular-eslint/eslint-plugin": "20.7.0",
+    "@angular-eslint/eslint-plugin-template": "20.7.0",
+    "@angular-eslint/schematics": "20.7.0",
+    "@angular-eslint/template-parser": "20.7.0",
+    "@angular/compiler-cli": "20.3.25",
+    "@angular/language-service": "20.3.25",
+    "@cspell/eslint-plugin": "^10.0.1",
+    "@nx/angular": "22.7.6",
+    "@nx/eslint-plugin": "22.5.4",
+    "@nx/workspace": "22.5.4",
+    "@playwright/test": "^1.61.1",
+    "@reportportal/agent-js-playwright": "^5.4.0",
     "@types/event-emitter": "^0.3.3",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "18.19.9",
+    "@types/node": "25.0.3",
     "@types/superagent": "^4.1.10",
-    "@typescript-eslint/eslint-plugin": "^7.2.0",
-    "@typescript-eslint/parser": "^7.2.0",
-    "@typescript-eslint/utils": "^8.40.0",
-    "axe-playwright": "^2.2.2",
-    "dotenv": "17.2.2",
-    "dotenv-expand": "^5.1.0",
+    "@typescript-eslint/eslint-plugin": "8.55.0",
+    "@typescript-eslint/parser": "8.55.0",
+    "@typescript-eslint/utils": "8.55.0",
+    "axe-playwright": "^1.2.3",
     "eslint": "^8.57.0",
     "eslint-plugin-ban": "^2.0.0",
     "eslint-plugin-import": "2.32.0",
     "eslint-plugin-jsdoc": "50.3.0",
     "eslint-plugin-license-header": "^0.8.0",
-    "eslint-plugin-playwright": "^1.6.2",
+    "eslint-plugin-playwright": "^2.4.0",
     "eslint-plugin-prefer-arrow": "1.2.3",
     "eslint-plugin-prettier": "5.5.4",
     "eslint-plugin-rxjs": "^5.0.2",
     "eslint-plugin-unicorn": "^54.0.0",
     "husky": "^9.0.11",
-    "jasmine-core": "~5.12.0",
+    "jasmine-core": "~6.0.1",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "^6.4.4",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "^2.2.0",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "^2.0.0",
+    "karma-jasmine-html-reporter": "^2.2.0",
     "karma-mocha-reporter": "^2.2.5",
-    "lint-staged": "^16.2.4",
-    "ng-packagr": "19.2.2",
+    "lint-staged": "^16.4.0",
+    "ng-packagr": "20.3.2",
     "node-stream-zip": "^1.14.0",
-    "nx": "21.5.2",
-    "prettier": "3.5.3",
-    "stylelint": "^15.11.0",
-    "stylelint-config-standard-scss": "^7.0.1",
+    "nx": "23.0.1",
+    "prettier": "3.8.1",
+    "stylelint": "^16.0.0",
+    "stylelint-config-standard-scss": "^13.0.0",
     "ts-node": "^10.9.2",
-    "typescript": "5.8.2",
-    "webpack-bundle-analyzer": "^4.10.2"
+    "typescript": "5.9.3",
+    "webpack-bundle-analyzer": "^5.3.0"
   }
 }
